+++ iocore/cluster/P_ClusterLib.h
@@ -39,7 +39,7 @@ extern void cluster_raise_priority(ClusterHandler *, ClusterVConnState *);
 extern void cluster_schedule(ClusterHandler *, ClusterVConnection *, ClusterVConnState *);
 extern void cluster_reschedule(ClusterHandler *, ClusterVConnection *, ClusterVConnState *);
 extern void cluster_disable(ClusterHandler *, ClusterVConnection *, ClusterVConnState *);
-extern void cluster_update_priority(ClusterHandler *, ClusterVConnection *, ClusterVConnState *, int, int);
+extern void cluster_update_priority(ClusterHandler *, ClusterVConnection *, ClusterVConnState *, int64_t, int64_t);
 extern void cluster_bump(ClusterHandler *, ClusterVConnectionBase *, ClusterVConnState *, int);
 extern int iov_memcpy(IOVec *, int, int, char *);
